# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Smart Pig Inventory Business API
  version: 0.4.0
  description: >-
    路线 B+ 的业务合同基线。所有写操作接受 X-Idempotency-Key，所有时间使用 ISO 8601；
    离线采集必须完成 package/blob/manifest/commit 后才创建正式推理任务。ROI 使用完成
    EXIF 方向纠正后的图像坐标，并归一化到 0..1。
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [Identity]
      security: []
      summary: 科研组账号登录
      operationId: login
      parameters:
        - {$ref: '#/components/parameters/IdempotencyKey'}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TokenPair'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
...
  /inventory-reports/daily:
    get:
      tags: [Inventory]
      summary: 查询指定业务日期已确认的日盘点
      operationId: getDailyInventoryReport
      parameters:
        - name: businessDate
          in: query
          required: true
          schema: {type: string, format: date}
      responses:
        '200':
          description: 只含已确认结果的日盘点报表
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/DailyInventoryRecord'}
  /inventory-reports/aggregate:
    get:
      tags: [Inventory]
      summary: 按栏舍和日期范围计算已确认日盘点的综合平均
      operationId: getAggregateInventoryReport
      parameters:
        - {$ref: '#/components/parameters/PenId'}
        - name: from
          in: query
          required: true
          schema: {type: string, format: date}
        - name: to
          in: query
          required: true
          schema: {type: string, format: date}
      responses:
        '200':
          description: 原始均值、展示整数和参与日期；无确认记录时展示数量为空
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AggregateInventoryReport'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    inferenceServiceKey:
      type: apiKey
      in: header
      name: X-Inference-Service-Key
  parameters:
    IdempotencyKey:
      name: X-Idempotency-Key
      in: header
      required: true
      schema: {type: string, format: uuid}
...
    PenId:
      name: penId
      in: query
      required: true
      schema: {type: string, format: uuid}
...
  responses:
    Unauthorized:
      description: 未提供有效身份、凭据错误或访问令牌已经失效
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    TooManyRequests:
      description: 请求过于频繁
      headers:
        Retry-After:
          schema: {type: integer, minimum: 1}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
  schemas:
...
    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password]
      properties:
        username: {type: string, minLength: 1, maxLength: 128}
        password: {type: string, minLength: 8, maxLength: 256, format: password}
...
    TokenPair:
      type: object
      additionalProperties: false
      required: [accessToken, refreshToken, tokenType, accessTokenExpiresAt, refreshTokenExpiresAt]
      properties:
        accessToken: {type: string, minLength: 1}
        refreshToken: {type: string, minLength: 32}
        tokenType: {type: string, const: Bearer}
        accessTokenExpiresAt: {type: string, format: date-time}
        refreshTokenExpiresAt: {type: string, format: date-time}
...
    DailyInventoryRecord:
      type: object
      required: [sessionId, penId, buildingCode, penCode, businessDate, confirmedCount]
      properties:
        sessionId: {type: string, format: uuid}
        penId: {type: string, format: uuid}
        buildingCode: {type: string}
        buildingName: {type: string}
        penCode: {type: string}
        penName: {type: string}
        businessDate: {type: string, format: date}
        confirmedCount: {type: integer, minimum: 0}
    AggregateInventoryReport:
      type: object
      required: [penId, from, to, rawMean, roundedCount, includedDates]
      properties:
        penId: {type: string, format: uuid}
        from: {type: string, format: date}
        to: {type: string, format: date}
        rawMean: {type: [number, 'null'], multipleOf: 0.0001}
        roundedCount: {type: [integer, 'null'], minimum: 0}
        includedDates:
          type: array
          items: {type: string, format: date}
...
    Problem:
      type: object
      required: [type, title, status, correlationId]
      properties:
        type: {type: string, format: uri-reference}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        code: {type: string}
        correlationId: {type: string}